Resource is destroying contents server.conf on plugin run
This should resolve the issue where the /etc/openvpn/server.conf
file is being overwritten and cleared of all config params save
for the requirements of that this cookbook does of adding the
`plugin` and `tmp-dir` params.

Additional work may be required to resolve this.  When calling
this resource in a different cookbook and you attempt to converge
it with Ubuntu 14.04 and Chef 12/14, it succeeds without issue.
However, when building against Ubuntu 18.04 and Chef 14, it will
fail.

The behavior is unexpected.  The reason for the failure occurs due
to the openvpn server.conf file containing references to the `plugin`,
which gets injected before the cookbook has the opportunity to
install the appropriate okta plugin files.  This leads to openvpn
failing to start due to errors of missing files, the okta plugin
files.

I was hoping that `edit_resource` would only modify the conf
resource AFTER all of openvpn was installed and configured.

Because chefspec doesn't respect node attributes,
and it's not the best practice to use them within a resource.  we
will modify the `openvpn_conf` resource, injecting the plugin and
tmp-dir params.  Then, after this custom resource has ran, it will
override the node['openvpn']['config'] attributes to include
['plugin'] and ['tmp-dir'].

We've also cleaned up some methods to provide a cleaner
separation of additional params we are adding.